clean migration files
# Conflicts:
#	speakwise/speakwise/speakwise/organizers/models.py
#	speakwise/speakwise/speakwise/speakers/admin.py
