+++ speakwise/speakwise/config/urls.py
@@ -48,9 +48,7 @@ urlpatterns += [
 ]
 
 # load media files
-urlpatterns += static(
-    settings.MEDIA_URL, document_root=settings.MEDIA_ROOT
-)
+urlpatterns += static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
 
 if settings.DEBUG:
     """This allows the error pages to be debugged during development, just visit
