+++ speakwise/speakwise/speakwise/events/serializers_extended.py
@@ -2,22 +2,6 @@
 
 from rest_framework import serializers
 from .serializers import EventSerializer, SessionSerializer
-from speakwise.speakers.serializers import SpeakerProfileSerializer
 
 
 # is this really necessary? I think we can add this to the serializers file and inherit from the EventSerializer as we did here.
-class EventWithGuestSpeakersSerializer(EventSerializer):
-    """Extended Event serializer that includes full speaker profile data."""
-
-    speaker_profiles = serializers.SerializerMethodField()
-    event_sessions = serializers.SerializerMethodField()
-
-    def get_speaker_profiles(self, obj):
-        """Get detailed speaker profiles for this event."""
-        speakers = obj.speakers.all()
-        return SpeakerProfileSerializer(speakers, many=True).data
-
-    def get_event_sessions(self, obj):
-        """Get sessions for this event with speaker details."""
-        sessions = obj.sessions.all()
-        return SessionSerializer(sessions, many=True).data
