+++ speakwise/speakwise/speakwise/talks/admin.py
@@ -6,9 +6,10 @@ from speakwise.talks.models import Talks
 
 class TalksAdmin(admin.ModelAdmin):
     """Admin view for the Talks model."""
-    list_display = ("event_id", "title","start_time", "end_time")
-    search_fields = ('title', 'event_id__title')
-    list_filter = ('event_id',)
+
+    list_display = ("event_id", "title", "start_time", "end_time")
+    search_fields = ("title", "event_id__title")
+    list_filter = ("event_id",)
     ordering = ("-created_at",)
